use email components
